fix: resolve mypy no-any-return errors in 5 modules
All 11 errors were "Returning Any from function declared to return float
or ndarray" caused by numpy/Python arithmetic expressions (** 0.25,
** (1/3), .copy()) being inferred as Any by mypy.

Fixes:
- stellar_wind.py:73 — wrap velocity() return in float()
- orbital.py:51     — annotate _separation as float, wrap Kepler
                       computation in float()
- orbital.py:146    — float() cast on orbital_velocity() return
- orbital.py:157    — float() cast on roche_lobe_radius() return
- _genesis_stubs.py:143 — float() cast on CREPTensor.Gamma property
- jet.py:168        — annotate _rodrigues_rotate result as np.ndarray
- jet.py:199,206,211 — annotate .copy() results as np.ndarray before return
- system.py:353     — annotate get_crep_state() result as dict[str, float]

+++ cygnus_jet_utac/_genesis_stubs.py
@@ -140,8 +140,10 @@ class CREPTensor:
     @property
     def Gamma(self) -> float:
         """Γ = (C · R · E · P)^(1/4) — aggregate CREP value."""
-        return (max(0.0, self.C) * max(0.0, self.R)
-                * max(0.0, self.E) * max(0.0, self.P)) ** 0.25
+        return float(
+            (max(0.0, self.C) * max(0.0, self.R)
+             * max(0.0, self.E) * max(0.0, self.P)) ** 0.25
+        )
 
     def to_dict(self) -> dict[str, float]:
         """Serialize to dictionary for logging and Zenodo export."""
